--- src/components/pages/home-page.tsx
@@ -1,8 +1,0 @@
-import { FC } from 'react';
-
-/**
- * Home page component.
- */
-export const HomePage: FC = () => (
-  <p>Hello World</p>
-);
