+++ src/pages/home-page.tsx
@@ -1,0 +1,12 @@
+import { FC } from 'react';
+import { Button } from '../components/button';
+
+/**
+ * Home page component.
+ */
+export const HomePage: FC = () => (
+  <div>
+    <p>Hello World</p>
+    <Button variant="error">Im a fucking button</Button>
+  </div>
+);
